chore(auth): extract interface from OpaJwtPrincipal
This is needed for Java 8 to create the proxy.
Otherwise a public default constructor would be needed.
