+++ sda-commons-server-auth/src/main/java/org/sdase/commons/server/opa/OpaJwtPrincipal.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.security.Principal;
 import java.util.Map;
+import org.sdase.commons.server.opa.internal.OpaJwtPrincipalImpl;
 
 /**
  * Principal for @{@link javax.ws.rs.core.SecurityContext} that optionally contains a JWT and a set
@@ -14,29 +15,7 @@ import java.util.Map;
  * javax.ws.rs.core.Context} when the {@link OpaBundle} is used to setup the open policy agent
  * configuration.
  */
-public class OpaJwtPrincipal implements Principal {
-
-  private static final String DEFAULT_NAME = OpaJwtPrincipal.class.getSimpleName();
-
-  private String name;
-  private String jwt;
-  private Map<String, Claim> claims;
-  private JsonNode constraints;
-  private ObjectMapper om;
-
-  @SuppressWarnings("unused")
-  OpaJwtPrincipal() {
-    // for proxy
-  }
-
-  private OpaJwtPrincipal(
-      String name, String jwt, Map<String, Claim> claims, JsonNode constraints, ObjectMapper om) {
-    this.name = name;
-    this.jwt = jwt;
-    this.claims = claims;
-    this.constraints = constraints;
-    this.om = om;
-  }
+public interface OpaJwtPrincipal extends Principal {
 
   /**
    * @param jwt The token this Principal is created from. May be required to pass it to other
@@ -46,30 +25,20 @@ public class OpaJwtPrincipal implements Principal {
    * @param constraints Authorization details used within the service for limiting result data
    * @return the principal that contains a jwt token, claims, and constraints that can be decoded
    */
-  public static OpaJwtPrincipal create(
+  static OpaJwtPrincipalImpl create(
       String jwt, Map<String, Claim> claims, JsonNode constraints, ObjectMapper om) {
-    return new OpaJwtPrincipal(DEFAULT_NAME, jwt, claims, constraints, om);
-  }
-
-  @Override
-  public String getName() {
-    return name;
+    String defaultName = OpaJwtPrincipal.class.getSimpleName();
+    return new OpaJwtPrincipalImpl(defaultName, jwt, claims, constraints, om);
   }
 
   /** @return the JWT as string */
-  public String getJwt() {
-    return jwt;
-  }
+  String getJwt();
 
   /** @return map with the claims decoded from the JWT */
-  public Map<String, Claim> getClaims() {
-    return claims;
-  }
+  Map<String, Claim> getClaims();
 
   /** @return the constraint object as JSON String */
-  public String getConstraints() {
-    return constraints.toString();
-  }
+  String getConstraints();
 
   /**
    * returns the constraint as Object. The object type must match the response from OPA sidecar
@@ -78,11 +47,5 @@ public class OpaJwtPrincipal implements Principal {
    * @param <T> type for correct casting
    * @return the object or null if no constraint exists
    */
-  public <T> T getConstraintsAsEntity(Class<T> resultType) {
-    if (constraints != null) {
-      return om.convertValue(constraints, resultType);
-    } else {
-      return null;
-    }
-  }
+  <T> T getConstraintsAsEntity(Class<T> resultType);
 }
